pyproject.toml
[project]
name = "aicalendarplanner"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrii Smidonov",email = "yuk0.terraf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph (>=0.3.2,<0.4.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "langchain-anthropic (>=0.3.8,<0.4.0)",
    "langsmith (>=0.3.11,<0.4.0)",
    "python-telegram-bot (>=21.11.1,<22.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "google-api-python-client (>=2.162.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "ruff (>=0.9.9,<0.10.0)",
    "ipython (>=9.0.1,<10.0.0)",
    "pydantic (>=2.10.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
